fix: terms and conditions copy
