+++ src/Screens/Flows/WalletCreation/WalletSetupScreen/WalletSetupScreen.tsx
@@ -14,7 +14,6 @@ import { useAnalyticTracking, useBottomSheetModal, useTheme } from "~Hooks"
 import { ImportWalletBottomSheet } from "./components"
 import { WalletSetupSvg } from "~Assets"
 import { AnalyticsEvent } from "~Constants"
-import { Linking } from "react-native"
 
 export const WalletSetupScreen = () => {
     const nav = useNavigation()
@@ -33,16 +32,6 @@ export const WalletSetupScreen = () => {
         track(AnalyticsEvent.SELECT_WALLET_IMPORT_WALLET)
         onOpen()
     }, [onOpen, track])
-
-    const goToTermsAndConditions = useCallback(() => {
-        const url = process.env.REACT_APP_TERMS_OF_SERVICE_URL
-        url && Linking.openURL(url)
-    }, [])
-
-    const goToPrivacyPolicy = useCallback(() => {
-        const url = process.env.REACT_APP_PRIVACY_POLICY_URL
-        url && Linking.openURL(url)
-    }, [])
 
     useEffect(() => {
         track(AnalyticsEvent.PAGE_LOADED_IMPORT_OR_CREATE)
@@ -131,35 +120,7 @@ export const WalletSetupScreen = () => {
                                 color={theme.colors.text}
                             />
                         </BaseTouchableBox>
-                        <BaseSpacer height={24} />
-                        <BaseView
-                            alignSelf="center"
-                            flexDirection="row"
-                            justifyContent="center"
-                            alignItems="center"
-                            flexWrap="wrap">
-                            <BaseText typographyFont="body" align="center">
-                                {LL.BD_CREATE_WALLET_TYPE_USER_ACCEPTS()}
-                            </BaseText>
-                            <BaseText
-                                typographyFont="bodyMedium"
-                                underline
-                                align="center"
-                                onPress={goToTermsAndConditions}>
-                                {LL.COMMON_LBL_TERMS_AND_CONDITIONS()}
-                            </BaseText>
-                            <BaseText typographyFont="body" align="center">
-                                {" "}
-                                {LL.COMMON_LBL_AND()}{" "}
-                            </BaseText>
-                            <BaseText
-                                typographyFont="bodyMedium"
-                                underline
-                                align="center"
-                                onPress={goToPrivacyPolicy}>
-                                {LL.COMMON_LBL_PRIVACY_POLICY()}
-                            </BaseText>
-                        </BaseView>
+                        <BaseSpacer height={16} />
                     </BaseView>
 
                     <ImportWalletBottomSheet ref={ref} onClose={onClose} />
